% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{generate_design_matrix.xrf}
\alias{generate_design_matrix.xrf}
\title{Generate the design matrix from a RuleFit xrf model}
\usage{
\method{generate_design_matrix}{xrf}(object, newdata, sparse = TRUE)
}
\arguments{
\item{object}{an object of class xrf}

\item{sparse}{a logical indicating whether a sparse design matrix should be used}
}
\description{
Generate the design matrix from a RuleFit xrf model
}
\author{
yama1968
}
