% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{generate_xrf_design_matrix}
\alias{generate_xrf_design_matrix}
\title{Generate the design matrix from an eXtreme RuleFit model}
\usage{
generate_xrf_design_matrix(object, newdata, sparse = TRUE)
}
\arguments{
\item{object}{an object of class xrf}

\item{newdata}{data to generate on}

\item{sparse}{a logical indicating whether a sparse design matrix should be used}
}
\description{
Generate the design matrix from an eXtreme RuleFit model
}
