% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnot.R
\name{glmnot.formula}
\alias{glmnot.formula}
\title{Elastic net regularization}
\usage{
\method{glmnot}{formula}(formula, data, family, alpha = 1, nfolds = 5,
  type.measure = "auc", pmax = +Inf, sparse = TRUE,
  weights = rep(1, nrow(data)), intercept = TRUE)
}
\arguments{
\item{formula}{a specification of the features used to build the design matrix}

\item{data}{a data frame used to build the model}

\item{family}{family of the fitted model}

\item{alpha}{the elastic net parameter}

\item{sparse}{logical indicating if the constructed design matrix should be sparse or dense}

\item{weights}{weights on observations in the train data}
}
\description{
Elastic net regularization
}
\author{
kholub
}
