% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnot.R
\name{glmnot}
\alias{glmnot}
\title{Elastic net regularized regression}
\usage{
glmnot(object, ...)
}
\description{
S3 method for elastic net fits - a wrapper on glmnet with some quality of life improvements.
See \code{\link{glmnot.formula}} for preferred entry point
}
\author{
kholub
}
