% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnot.R
\name{plot.glmnot}
\alias{plot.glmnot}
\title{Produces regularization path plot (delegated to optimal glmnet fit)}
\usage{
\method{plot}{glmnot}(object, ...)
}
\description{
Produces regularization path plot (delegated to optimal glmnet fit)
}
\author{
kholub
}
