% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnot.R
\name{predict.glmnot}
\alias{predict.glmnot}
\title{Delegates a cv.glmnet}
\usage{
\method{predict}{glmnot}(object, newdata, sparse = TRUE,
  lambda = "lambda.min", type = "response")
}
\arguments{
\item{object}{a glmnnot object to predict from}

\item{newdata}{a matrix or dataframe to draw predictions from, dependent on how the model was trained}

\item{sparse}{if true and supplied data is a dataframe, the constructed design matrix is sparse}
}
\description{
Accepts a data frame or design matrix, dependent on what the model was trained with. 
If a model was trained with a data.frame, it can be predicted with either a matrix or a data.frame
}
\examples{
dat <- data.frame(
  x = c(1,2,3,4,55,6,2,3, 1,2,3),
  z = as.factor(c('a','b','c','a','c','b','c', 'c', 'a', 'b', 'c')),
  y = c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, FALSE, FALSE, TRUE, TRUE, TRUE))
test_dat <- data.frame(x = c(10,5), z = c('c', 'd'))
predict(m, newdata = test_dat)

}
\author{
kholub
}
