% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{predict.xrf}
\alias{predict.xrf}
\title{Draw predictions from a RuleFit model}
\usage{
\method{predict}{xrf}(object, newdata, sparse = "TRUE",
  lambda = "lambda.min", type = "response")
}
\description{
Draw predictions from a RuleFit model
}
\author{
kholub
}
