% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{print.xrf}
\alias{print.xrf}
\title{Print an eXtreme RuleFit model}
\usage{
\method{print}{xrf}(x, ...)
}
\arguments{
\item{x}{an xrf object to be printed}

\item{...}{ignored arguments}
}
\description{
Print an eXtreme RuleFit model
}
