% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_XRF.R
\name{read_XRF}
\alias{read_XRF}
\title{import xrf data}
\usage{
read_XRF(datapath, infopath, setuppath, year)
}
\arguments{
\item{datapath}{name of my XRFdata.TXT file}

\item{infopath}{name of my Infofile.xlsx file}

\item{setuppath}{name of xrf_setup.xlsx file}

\item{year}{year the drift was measured}
}
\description{
import xrf data
}
\examples{
#\dontrun {
 # read_XRF(datapath = "data/Test.TXT",  infopath = "data/Infofile.xlsx",
  #         setuppath = "data/xrf_setup.xlsx", year = "2019")}

# Example with external data
datadir <- system.file("extdata", package = "XRF")
read_XRF(datapath = file.path(datadir, "xrf_rawdata.TXT"),
           infopath = file.path(datadir,"project_info.xlsx"),
           setuppath = file.path(datadir,"xrf_setup.xlsx"),
           year = "2019")
}
