% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{summary.xrf}
\alias{summary.xrf}
\title{Summarize an eXtreme RuleFit model}
\usage{
\method{summary}{xrf}(object, ...)
}
\arguments{
\item{object}{an object of class xrf}
}
\description{
Summarize an eXtreme RuleFit model
}
\author{
kholub
}
