% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{xrf.formula}
\alias{xrf.formula}
\title{Fit a RuleFit model}
\usage{
\method{xrf}{formula}(object, data, family, xgb_control = list(nrounds = 100),
  glm_control = list(type.measure = "deviance", nfolds = 5),
  max_rule_correlation = 0.99, sparse = TRUE)
}
\description{
See Friedman & Popescu (2008) for a description of the general RuleFit algorithm.
This method uses XGBoost to fit a tree ensemble, extracts a ruleset as the conjunction of tree 
traversals, and fits a sparse linear model to the resulting feature set
(including the original feature set) using glmnet.
}
\examples{
m <- xrf(Petal.Length ~ ., iris, family = 'gaussian')

}
\references{
Friedman, J. H., & Popescu, B. E. (2008). Predictive learning via rule 
ensembles. \emph{The Annals of Applied Statistics, 2}(3), 916-954.
}
\author{
kholub
}
