% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{xrf}
\alias{xrf}
\title{Fit an eXtreme RuleFit model}
\usage{
xrf(object, ...)
}
\arguments{
\item{object}{an object describing the model to be fit}

\item{...}{additional arguments}
}
\description{
S3 method for building an "eXtreme RuleFit" model.
See \code{\link{xrf.formula}} for preferred entry point
}
