% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{coef.xrf}
\alias{coef.xrf}
\title{Produce rules & coefficients for the RuleFit model}
\usage{
\method{coef}{xrf}(object, lambda = "lambda.min", ...)
}
\arguments{
\item{object}{an object of class xrf}

\item{lambda}{the lasso penalty parameter to be applied}
}
\description{
Produce rules & coefficients for the RuleFit model
}
\author{
kholub
}
