% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnot.R
\name{glmnot.default}
\alias{glmnot.default}
\title{Elastic net regularization algorithm}
\usage{
\method{glmnot}{default}(X, y, family, alpha = 1, nfolds = 5,
  type.measure = "auc", formula = NULL, xlev = NULL,
  weights = rep(1, nrow(X)))
}
\arguments{
\item{X}{a design matrix (sparse or otherwise) consisting of predictors to train on}

\item{y}{a vector of responses to train on}

\item{family}{the family of the model to be fitted}

\item{alpha}{the elastic net parameter}

\item{formula}{the formula specify the design matrix (optional)}

\item{xlev}{per feature levels (optional)}

\item{weights}{weights on observations in X}
}
\description{
Elastic net regularization algorithm
}
\author{
kholub
}
