% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnot.R
\name{glmnot.formula}
\alias{glmnot.formula}
\title{Elastic net regularization}
\usage{
\method{glmnot}{formula}(formula, data, family, alpha = 1,
  type.measure = "auc", sparse = TRUE, glm_control)
}
\arguments{
\item{formula}{a specification of the features used to build the design matrix}

\item{data}{a data frame used to build the model}

\item{family}{family of the fitted model}

\item{alpha}{the elastic net parameter}

\item{glm_control}{other parameters to glm}

\item{formula}{the formula specify the design matrix (optional)}

\item{xlev}{per feature levels (optional)}

\item{weights}{weights on observations in X}
}
\description{
Elastic net regularization
}
\author{
kholub
}
