% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script.R
\name{read_XRF}
\alias{read_XRF}
\title{import xrf data}
\usage{
read_XRF(datapath, infopath, setuppath, year)
}
\arguments{
\item{datapath}{name of my XRFdata.TXT file}

\item{infopath}{name of my Infofile.xlsx file}

\item{setuppath}{name of xrf_setup.xlsx file}

\item{year}{yeat the drift was measured}
}
\description{
import xrf data
}
\examples{
\dontrun {read_XRF(datapath = "data/Test.TXT",  infopath = "data/Infofile.xlsx", setuppath = "data/xrf_setup.xlsx", year = "2019")
}
}
