% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{summary.xrf}
\alias{summary.xrf}
\title{Summarize an eXtreme RuleFit model}
\usage{
\method{summary}{xrf}(object, ...)
}
\arguments{
\item{object}{an object of class "xrf"}

\item{...}{ignored arguments}
}
\description{
Summarize an eXtreme RuleFit model
}
\examples{
m <- xrf(Petal.Length ~ ., iris, 
         xgb_control = list(nrounds = 20, max_depth = 2),
         family = 'gaussian')
summary(m)

}
