% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrf.R
\name{xrf}
\alias{xrf}
\title{Fit a RuleFit model}
\usage{
xrf(object, ...)
}
\description{
S3 method for building an "eXtreme RuleFit" model.
See \code{\link{xrf.formula}} for preferred entry point
}
\author{
kholub
}
